+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="/css/basestyle.css">
+    <link rel="stylesheet" href="css/basestyle.css">
 </head>
 <body>
     <div class="playzone" id="playzone">
@@ -17,5 +17,5 @@
     </div>
    
 </body>
-<script src="js/main.js"></script>
+<script src="js/main.js?v=6" type="module"></script>
 </html>
